% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcpbayes.R
\name{BIC.cpbayes}
\alias{BIC.cpbayes}
\title{BIC information criteria}
\usage{
\method{BIC}{cpbayes}(object, ...)
}
\arguments{
\item{object}{cpbayes object}

\item{...}{additional arguments}
}
\description{
BIC information criteria
}
