% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/compois.R
\name{rcompois}
\alias{rcompois}
\alias{dcompois}
\alias{pcompois}
\alias{qcompois}
\title{COM-Poisson distribution}
\usage{
rcompois(n, mu, nu)

dcompois(y, mu, nu, log = FALSE)

pcompois(q, mu, nu, lower.tail = TRUE, log.p = FALSE)

qcompois(p, mu, nu)
}
\arguments{
\item{n}{Integer. Number of observations.}

\item{mu}{Numeric value. Location parameter, \eqn{\mu >0}.}

\item{nu}{Numeric value. Dispersion parameter, \eqn{\nu >0}.}

\item{y}{Count vector.}

\item{log}{Logical; if TRUE, returns the log-probabilities.}

\item{q}{Numeric value. Quantile of probability mass function.}

\item{lower.tail}{Logical; if TRUE \eqn{P(X >x)}, otherwise \eqn{P(X \le x)}.}

\item{log.p}{Logical; if TRUE, returns the log-probabilities.}

\item{p}{Numeric value; probability.}
}
\value{
Integer count vector of \code{n} independent draws.
}
\description{
Exact sampling, approximate density, distribution and quantile computation.
}
\details{
The COM-Poisson distribution has density 
\deqn{f(y|\mu, \nu) = (\mu^y/y!)^\nu 1/Z(\mu, \nu) } 
for \eqn{y = 0, 1, 2, ...}, where \eqn{Z(\mu, \nu) = \sum_{y=0}^\infty (\mu^y/y!)^\nu }. 

If \eqn{\nu <1} the distribution is overdispersed (\eqn{E(Y) < Var(Y)}), exhibits underdispersion if \eqn{\nu >1} or reduces to a Poisson (equidispersed) if \eqn{\nu=1}.
The Geometric and Bernoulli distributions can be obtained as limiting cases. \cr 

Exact sampling \eqn{f(y|\mu, \nu)} is done with the \href{10.1214/20-BA1230}{fast-rejection sampler} sampler of Benson and Friel (2021),
for which C++ implementation is provided in \code{rcompois}.


Density, distribution and quantiles are unavailable from intractability of \eqn{Z(\mu, \nu)}
and can be approximated using a truncated sum \eqn{Z(\mu, \nu) = \sum_{y=0}^T (\mu^y/y!)^\nu }.
\code{dcompois}, \code{qcompois} and \code{pcompois} apply a \eqn{Z(\mu, \nu)} approximation that
recursively increments \eqn{T} until difference between successive probabilities is below \eqn{10^{-5}}. \cr
   
Asymptotic moments of the distribution ((Shmueli et al. (2005))) are \eqn{E(Y) ~ \mu + 1/(2\nu) -1/2}, \eqn{Var(Y) ~ \mu/\nu}. 

For further details on the fast-rejection sampler, see Benson and Friel (2021).
}
\examples{
y1 = rcompois(100, 1, 0.5) ## independent overdispersed counts
y2 = rcompois(100, 2, 1) ## Poisson random draws
y3 = rcompois(100, 0.5, 1.5) ## an underdispersed count vector

}
\references{
Benson, A. and Friel, N. (2021) Bayesian Inference, Model Selection and Likelihood Estimation using Fast Rejection Sampling: The Conway-Maxwell-Poisson Distribution.
Bayesian Analysis (16) 905-931.

Shmueli, G., Minka, T., Kadane, J, Borle, S. and Boatwright, P. (2005) A useful distribution for fitting discrete data: Revival of the Conway-Maxwell-Poisson distribution.
Journal of the Royal Statistical Society (C) (54) 127-142.
}
