% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exchange_noreg}
\alias{exchange_noreg}
\title{MCMC sampling: no regression case}
\usage{
exchange_noreg(y, params0, sigma, n_iter, burn_in, hyperparams)
}
\arguments{
\item{y}{data}

\item{params0}{initial parameter values}

\item{sigma}{c(sigma_mu, sigma_nu) vector of proposal parameters}

\item{n_iter}{number of iterations to store}

\item{burn_in}{burn-in period}

\item{hyperparams}{named list containing 'shape' c(shape_mu, shape_nu) gamma prior shape parameters and
'rate' c(rate_mu, rate_nu) Gamma prior rate parameters.}
}
\description{
MCMC sampling: no regression case
}
