% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcpbayes.R
\name{fitcpbayes_single}
\alias{fitcpbayes_single}
\title{Runs a single chain of COM-Poisson GLM}
\usage{
fitcpbayes_single(X_mu, X_nu, y, burnin, niter, prior_mu, prior_nu)
}
\arguments{
\item{X_mu}{parsed regression matrix for location}

\item{X_nu}{parsed regression matrix for dispersion}

\item{y}{parsed response variable}

\item{burnin}{integer. Number of iterations to discard as burn-in period.}

\item{niter}{integer. Number of iterations to store.}

\item{prior_mu}{optional named list with 'shape' and 'rate' if no regression on location, or 'mean' and 'sd' if running CP GLM. 
Defaults to ('shape' =2, 'rate' =2) or ('mean' =0, 'sd' =1).}

\item{prior_nu}{optional named list with 'shape' and 'rate' if no regression on dispersion, or 'mean' and 'sd' if running CP GLM.
Defaults to ('shape' =2, 'rate' =2) or ('mean' =0, 'sd' =1).}
}
\description{
Runs a single chain of COM-Poisson GLM
}
