% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\docType{data}
\name{inventory}
\alias{inventory}
\title{Inventory data}
\format{
An object of class \code{data.frame} with 3168 rows and 1 columns.
}
\source{
Shmueli et al. (2005)
}
\usage{
inventory
}
\description{
Quarterly sales counts of clothing items in 3168 stores.
}
\references{
Shmueli et al (2005). 
A useful distribution for fitting discrete data: Revival of the Conway-Maxwell-Poisson distribution. 
Journal of the Royal Statistical Society: Series C (Applied Statistics) 54(1), 127-142.
}
\keyword{datasets}
