% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcpbayes.R
\name{plot.cpbayes}
\alias{plot.cpbayes}
\title{Plot cpbayes MCMC output}
\usage{
\method{plot}{cpbayes}(x, type = "density", ...)
}
\arguments{
\item{x}{cpbayes object}

\item{type}{trace or density plot}

\item{...}{additional arguments}
}
\description{
Plot cpbayes MCMC output
}
