% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proposal_normal}
\alias{proposal_normal}
\title{Normal proposal}
\usage{
proposal_normal(index, param, sigma)
}
\arguments{
\item{index}{starting from 0, the parameter to propose}

\item{param}{vector}

\item{sigma}{vector}
}
\value{
param vector with updated position 'index'
}
\description{
Normal proposal
}
