% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompois}
\alias{rcompois}
\title{COM-Poisson rejection sampling}
\usage{
rcompois(n, mu, nu)
}
\arguments{
\item{n}{number of samples}

\item{mu}{location parameter}

\item{nu}{dispersion parameter}
}
\description{
COM-Poisson rejection sampling
}
