% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompoisreg}
\alias{rcompoisreg}
\title{COM-Poisson sampling under mu and nu vectors}
\usage{
rcompoisreg(mu, nu)
}
\arguments{
\item{mu}{vector from linking function exp(beta_{mu,0} + beta_{mu,1}*x1 + ...)}

\item{nu}{vector from linking function exp(beta_{nu,0} + beta_{nu,1}*x1 + ...)}
}
\description{
COM-Poisson sampling under mu and nu vectors
}
