% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcpbayes.R
\name{summary.cpbayes}
\alias{summary.cpbayes}
\title{Summarise cpbayes MCMC output}
\usage{
\method{summary}{cpbayes}(object, ...)
}
\arguments{
\item{object}{cpbayes object}

\item{...}{additional arguments}
}
\description{
Summarise cpbayes MCMC output
}
