% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_beta_nu}
\alias{update_beta_nu}
\title{Exchange move for dispersion regression parameters}
\usage{
update_beta_nu(
  index,
  beta_mu,
  beta_nu,
  y,
  X_mu,
  X_nu,
  sigma_nu,
  hyperparams_nu
)
}
\arguments{
\item{index}{index of parameter to update, from 0}

\item{beta_mu}{current beta_mu}

\item{beta_nu}{current beta_nu}

\item{y}{response vector}

\item{X_mu}{location model matrix}

\item{X_nu}{dispersion model matrix}

\item{sigma_nu}{proposal variances}

\item{hyperparams_nu}{named list with 'mean' and 'sd'}
}
\description{
Exchange move for dispersion regression parameters
}
