% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Zhat}
\alias{Zhat}
\title{Estimate COM-Poisson normalising constant with the fast-rejection sampler}
\usage{
Zhat(r, mu, nu)
}
\arguments{
\item{r}{number of fast-rejection sampler draws}

\item{mu}{location parameter}

\item{nu}{dispersion parameter}
}
\value{
double
}
\description{
Estimate COM-Poisson normalising constant with the fast-rejection sampler
}
