% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcpbayes.R
\name{summary.cpbayes}
\alias{summary.cpbayes}
\alias{plot.cpbayes}
\title{Summarising COM-Poisson GLM model fits}
\usage{
\method{summary}{cpbayes}(object, ...)

\method{plot}{cpbayes}(x, type = "density", ...)
}
\arguments{
\item{object}{cpbayes object}

\item{...}{additional arguments}

\item{x}{cpbayes object}

\item{type}{trace or density plot}
}
\description{
Plotting and summarising \code{methods} for \code{cpbayes} objects obtaining with \code{\link{fitcpbayes}}.
}
\details{
Summary: neff and Rhat...
Plot: ...
}
