% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcpbayes.R
\name{fitcpbayes}
\alias{fitcpbayes}
\title{Run COM-Poisson GLM}
\usage{
fitcpbayes(formula_beta, formula_nu, data, burnin, niter, prior_mu, prior_nu)
}
\arguments{
\item{formula_beta}{regression formula for location}

\item{formula_nu}{regression formula for dispersion}

\item{data}{data frame}

\item{burnin}{integer. Number of iterations to discard as burn-in period.}

\item{niter}{integer. Number of iterations to store.}

\item{prior_mu}{optional named list with 'shape' and 'rate' if no regression on location, or 'mean' and 'sd' if running CP GLM. 
Defaults to ('shape' =2, 'rate' =2) or ('mean' =0, 'sd' =1).}

\item{prior_nu}{optional named list with 'shape' and 'rate' if no regression on dispersion, or 'mean' and 'sd' if running CP GLM.
Defaults to ('shape' =2, 'rate' =2) or ('mean' =0, 'sd' =1).}
}
\description{
Run COM-Poisson GLM
}
