% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompoisreg}
\alias{rcompoisreg}
\title{COM-Poisson sampling under mu and nu vectors}
\usage{
rcompoisreg(beta_mu, beta_nu, X_mu, X_nu)
}
\arguments{
\item{beta_mu}{location regression parameters}

\item{beta_nu}{dispersion regression parameter}

\item{X_mu}{model matrix for mu (first column=1: intercept)}

\item{X_nu}{model matrix for nu (first column=1: intercept)}
}
\description{
COM-Poisson sampling under mu and nu vectors
}
