% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeoverbids.R
\docType{data}
\name{takeoverbids}
\alias{takeoverbids}
\title{Takeoverbids data}
\format{
\describe{
\item{numbids}{Count of takeover bids excluding the initial bid.}
\item{bidprem}{Bid price divided by price 14 working days before bid.}
\item{finrest}{Did management propose changes in ownership structure? (1=YES,0=NO)}
\item{insthold}{Percentage of stock held by institutions.}
\item{leglrest}{Indicator variable for legal defence by lawsuit.}
\item{rearest}{Did management propose changes in asset structure? (1=YES,0=NO)}
\item{regulatn}{Did federal regulators intervene? (1=YES,0=NO)}
\item{size}{Total book value of assets. (USD billions)}
\item{takeover}{Indicates if the company was being taken over.}
\item{weeks}{Time in weeks between the initial and final offers.}
\item{whtknght}{Did management invite friendly third-party bid?  (1=YES,0=NO)}
\item{sizesq}{Book value squared.}
}
}
\usage{
takeoverbids
}
\description{
Data set from Jaggia and Thosar (1993) on the number of bids received by U.S. firms that were
targets of bid offers during 1978-1985 and taken over within 1 year of the initial offer. 
The response variable is the number of takeover bids excluding the initial bid for a particular company.
Multiple covariates are available with descriptions below from Saez-Castillo & Conde-Sanchez, A. (2013).
}
\references{
Jaggia, S. and S. Thosar (1993). Multiple bids as a consequence of target management resistance: 
A count data approach. Review of Quantitative Finance and Accounting 3(4), 447-457.

Saez-Castillo, Antonio & Conde-Sanchez, A. (2013). A hyper-Poisson regression model for overdispersed 
and underdispersed count data. Computational Statistics & Data Analysis. 61. 148-157.
}
\keyword{datasets}
